Revert "merge ignition-gazebo6"
This reverts commit 4e92b3f989e34ad179d70db19caeb49276f214c4, reversing
changes made to 1f1e2b3f8fb64300d8641ce73a651705541d6b7f.

+++ src/systems/optical_tactile_plugin/Visualization.cc
@@ -170,7 +170,7 @@ void OpticalTactilePluginVisualization::InitializeNormalForcesMarkerMsgs(
   ignition::msgs::Set(_positionMarkerMsg.mutable_material()->
     mutable_diffuse(), math::Color(0, 0, 1, 1));
   _positionMarkerMsg.mutable_lifetime()->set_nsec(
-    static_cast<int>(this->cameraUpdateRate * 1000000000));
+    this->cameraUpdateRate * 1000000000);
 
   // Green lines for forces
   ignition::msgs::Set(_forceMarkerMsg.mutable_material()->
@@ -178,7 +178,7 @@ void OpticalTactilePluginVisualization::InitializeNormalForcesMarkerMsgs(
   ignition::msgs::Set(_forceMarkerMsg.mutable_material()->
     mutable_diffuse(), math::Color(0, 1, 0, 1));
   _forceMarkerMsg.mutable_lifetime()->set_sec(
-    static_cast<int>(this->cameraUpdateRate * 1000000000));
+    this->cameraUpdateRate * 1000000000);
 }
 
 //////////////////////////////////////////////////
